+++ sources/org/apache/batik/dom/svg/SVGOMEllipseElement.java
@@ -138,7 +138,7 @@ public class SVGOMEllipseElement
     /**
      * Gets how percentage values are interpreted by the given attribute.
      */
-    protected int getAttributePercentageInterpretation(String ns, String ln) {
+    protected short getAttributePercentageInterpretation(String ns, String ln) {
         if (ns == null) {
             if (ln.equals(SVG_CX_ATTRIBUTE) || ln.equals(SVG_RX_ATTRIBUTE)) {
                 return AnimationTarget.PERCENTAGE_VIEWPORT_WIDTH;
